Refactored tests and re-made the repositories smoke tests so that it will find the proper tests
