--- CMSforce/src/test/java/com/revature/cmsforce/CmSforceApplicationTests.java
@@ -1,18 +1,0 @@
-package com.revature.cmsforce;
-
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-@ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes=com.revature.cmsforce.CMSforceApplication.class)
-@SpringBootTest
-public class CmSforceApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
