+++ CMSforce/src/test/java/com/revature/servicestests/ModuleServiceTest.java
@@ -1,4 +1,4 @@
-package com.revature.services;
+package com.revature.servicestests;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -16,6 +16,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 
 import com.revature.entities.Content;
 import com.revature.entities.Module;
+import com.revature.services.ModuleService;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(classes=com.revature.cmsforce.CMSforceApplication.class)
